# Conflicts:
#	vispy/gloo/buffer.py

+++ examples/demo/gloo/imshow_cuts.py
@@ -30,21 +30,21 @@ values = np.linspace(0, 1, 512)[1:-1]
 colormaps[0, 0] = 0, 0, 1, 1  # Low values  (< vmin)
 colormaps[0, -1] = 0, 1, 0, 1  # High values (> vmax)
 colormaps[0, 1:-1, 0] = np.interp(values, [0.00, 0.33, 0.66, 1.00],
-                                          [0.00, 1.00, 1.00, 1.00])
+                                  [0.00, 1.00, 1.00, 1.00])
 colormaps[0, 1:-1, 1] = np.interp(values, [0.00, 0.33, 0.66, 1.00],
-                                          [0.00, 0.00, 1.00, 1.00])
+                                  [0.00, 0.00, 1.00, 1.00])
 colormaps[0, 1:-1, 2] = np.interp(values, [0.00, 0.33, 0.66, 1.00],
-                                          [0.00, 0.00, 0.00, 1.00])
+                                  [0.00, 0.00, 0.00, 1.00])
 
 # Grey colormap
 colormaps[1, 0] = 0, 0, 1, 1  # Low values (< vmin)
 colormaps[1, -1] = 0, 1, 0, 1  # High values (> vmax)
 colormaps[1, 1:-1, 0] = np.interp(values, [0.00, 1.00],
-                                          [0.00, 1.00])
+                                  [0.00, 1.00])
 colormaps[1, 1:-1, 1] = np.interp(values, [0.00, 1.00],
-                                          [0.00, 1.00])
+                                  [0.00, 1.00])
 colormaps[1, 1:-1, 2] = np.interp(values, [0.00, 1.00],
-                                          [0.00, 1.00])
+                                  [0.00, 1.00])
 # Jet colormap
 # ...
 
@@ -160,8 +160,8 @@ class Canvas(app.Canvas):
         yf = 1 - y/(h/2.)
         xf = x/(w/2.) - 1
 
-        x_norm = (x*512)//w
-        y_norm = (y*512)//h
+        x_norm = int((x*512)//w)
+        y_norm = int((y*512)//h)
 
         P = np.zeros((4+4+514+514, 2), np.float32)
